requirements.txt
Django>=5.0
openai>=1.52.0
python-dotenv>=1.0
Pillow>=10.0